When the stream is closed, close the database objects
The StreamingQueryRunner creates a DatabaseConnection,
PreparedStatement, and a ResultSet. None of those objects were being
closed.

This commit ensures they are closed by creating a Proxy of the
Stream. It just delegates all methods to the real stream, but when the
close method is invoked it closes the three database objects as well.