Refactor some tests, add new tests, and add public query that takes a Connection
The ResultSet handler subclasses have a single method that converts
the current row of a ResultSet into some other object. That is really
the only thing that should be tested.

The other methods in those tests were more like tests of the
StreamingQueryRunner.  So I moveed the test from the
ArrayStreamingHandler test there, then parameterized them so the same
tests could also be used for the MapStreamingHandler.

I also added a method to the StreamingQueryRunner that allows users to
pass it a connection to use. In that case, closing the stream only
closes the ResultSet and PreparedStatement, not the Connection.

Tests were also added to ensure that the ResultSet, PreparedStatement,
and Connection are closed when the stream is closed.