Deal with exceptions
When processing a row in a ResultSet, we may get an SQLException, or a
user's custom handler could throw any kind of RuntimeException.  There
is no way for the user to close the DB objects Connection,
PreparedStatement, or ResultSet, in that case, even if they know to
catch the RuntimeException.

This commit adds code to clean up those resources when there is an
exception processing the stream.